Splits TPCH benchmarking and unit tests
* Refactors common data generation utilities into a separate module in benchmarking/tpch/data_generation.py
* Refactors TPC-H Daft answers into a separate module in benchmarking/tpch/answers.py
* Creates an entrypoint for benchmarking in benchmarking/tpch/__main__.py
* Shares code between benchmarking entrypoint and TPC-H unit tests
